+++ cli/commands/root.go
@@ -2,9 +2,7 @@ package commands
 
 import (
 	"errors"
-	"fmt"
 	"io/fs"
-	"net/http"
 	"os"
 	"path/filepath"
 	"time"
@@ -16,13 +14,11 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/wundergraph/wundergraph/cli/helpers"
-	"github.com/wundergraph/wundergraph/pkg/cli/auth"
 	"github.com/wundergraph/wundergraph/pkg/config"
 	"github.com/wundergraph/wundergraph/pkg/files"
 	"github.com/wundergraph/wundergraph/pkg/logging"
-	"github.com/wundergraph/wundergraph/pkg/manifest"
 	"github.com/wundergraph/wundergraph/pkg/node"
-	"github.com/wundergraph/wundergraph/pkg/v2wundergraphapi"
+	"github.com/wundergraph/wundergraph/pkg/telemetry"
 )
 
 const (
@@ -38,9 +34,11 @@ const (
 var (
 	BuildInfo             node.BuildInfo
 	GitHubAuthDemo        node.GitHubAuthDemo
+	TelemetryClient       telemetry.Client
 	DotEnvFile            string
 	log                   *zap.Logger
 	serviceToken          string
+	cmdDurationMetric     telemetry.DurationMetric
 	_wunderGraphDirConfig string
 	disableCache          bool
 	clearCache            bool
@@ -59,15 +57,10 @@ var (
 var rootCmd = &cobra.Command{
 	Use:   "wunderctl",
 	Short: "wunderctl is the cli to manage, build and debug your WunderGraph applications",
-	Long: `wunderctl is the cli to manage, build and debug your WunderGraph applications
-
-Simply running "wunderctl" will check the wundergraph.manifest.json in the current directory and install all dependencies. 
-
-wunderctl is gathering anonymous usage data so that we can better understand how it's being used and improve it.
-You can opt out of this by setting the following environment variable: WUNDERGRAPH_DISABLE_METRICS
-`,
+	Long:  `wunderctl is the cli to manage, build and debug your WunderGraph applications.`,
+	// Don't show usage on error
+	SilenceUsage: true,
 	PersistentPreRunE: func(cmd *cobra.Command, args []string) error {
-
 		switch cmd.Name() {
 		// skip any setup to avoid logging anything
 		// because the command output data on stdout
@@ -99,8 +92,9 @@ You can opt out of this by setting the following environment variable: WUNDERGRA
 			if _, ok := err.(*fs.PathError); ok {
 				log.Debug("starting without env file")
 			} else {
-				log.Fatal("error loading env file",
+				log.Error("error loading env file",
 					zap.Error(err))
+				return err
 			}
 		} else {
 			log.Debug("env file successfully loaded",
@@ -119,41 +113,43 @@ You can opt out of this by setting the following environment variable: WUNDERGRA
 			}
 		}
 
+		// Check if we want to track telemetry for this command
+		if rootFlags.Telemetry && cmd.Annotations["telemetry"] == "true" {
+			TelemetryClient = telemetry.NewClient(
+				viper.GetString("API_URL"),
+				telemetry.MetricClientInfo{
+					WunderctlVersion: BuildInfo.Version,
+					IsCI:             os.Getenv("CI") != "" || os.Getenv("ci") != "",
+					AnonymousID:      viper.GetString("anonymousid"),
+				},
+				telemetry.WithTimeout(3*time.Second),
+				telemetry.WithLogger(log),
+				telemetry.WithDebug(rootFlags.TelemetryDebugMode),
+			)
+
+			cmdMetricName := telemetry.CobraFullCommandPathMetricName(cmd)
+
+			metricDurationName := telemetry.DurationMetricSuffix(cmdMetricName)
+			cmdDurationMetric = telemetry.NewDurationMetric(metricDurationName)
+
+			metricUsageName := telemetry.UsageMetricSuffix(cmdMetricName)
+			cmdUsageMetric := telemetry.NewUsageMetric(metricUsageName)
+
+			err := TelemetryClient.Send([]telemetry.Metric{cmdUsageMetric})
+
+			// AddMetric the usage of the command immediately
+			if rootFlags.TelemetryDebugMode {
+				if err != nil {
+					log.Error("Could not send telemetry data", zap.Error(err))
+				} else {
+					log.Info("Telemetry data sent")
+				}
+			}
+
+		}
+
 		return nil
 	},
-	RunE: func(cmd *cobra.Command, args []string) error {
-		client := InitWunderGraphApiClient()
-		wunderGraphDir, err := files.FindWunderGraphDir(_wunderGraphDirConfig)
-		if err != nil {
-			return err
-		}
-		man := manifest.New(log, client, wunderGraphDir)
-		err = man.Load()
-		if err != nil {
-			return fmt.Errorf("unable to load wundergraph.manifest.json")
-		}
-		err = man.PersistChanges()
-		if err != nil {
-			return fmt.Errorf("unable to persist manifest changes")
-		}
-		err = man.WriteIntegrationsFile()
-		if err != nil {
-			return fmt.Errorf("unable to write integrations file")
-		}
-		_, _ = green.Printf("API dependencies updated\n")
-		return nil
-	},
-}
-
-func InitWunderGraphApiClient() *v2wundergraphapi.Client {
-	token := authenticator().LoadRefreshAccessToken()
-	return InitWunderGraphApiClientWithToken(token)
-}
-
-func InitWunderGraphApiClientWithToken(token string) *v2wundergraphapi.Client {
-	return v2wundergraphapi.New(token, viper.GetString("API_URL"), &http.Client{
-		Timeout: time.Second * 10,
-	}, log)
 }
 
 type BuildTimeConfig struct {
@@ -163,11 +159,34 @@ type BuildTimeConfig struct {
 // Execute adds all child commands to the root command and sets flags appropriately.
 // This is called by main.main(). It only needs to happen once to the rootCmd.
 func Execute(buildInfo node.BuildInfo, githubAuthDemo node.GitHubAuthDemo) {
+	var err error
+
+	defer func() {
+		// In case of a panic or error we want to flush the telemetry data
+		if r := recover(); r != nil || err != nil {
+			FlushTelemetry()
+			os.Exit(1)
+		} else {
+			FlushTelemetry()
+			os.Exit(0)
+		}
+	}()
+
 	BuildInfo = buildInfo
 	GitHubAuthDemo = githubAuthDemo
-	if err := rootCmd.Execute(); err != nil {
-		fmt.Println(err)
-		os.Exit(1)
+	err = rootCmd.Execute()
+}
+
+func FlushTelemetry() {
+	if TelemetryClient != nil && rootFlags.Telemetry && cmdDurationMetric != nil {
+		err := TelemetryClient.Send([]telemetry.Metric{cmdDurationMetric()})
+		if rootFlags.TelemetryDebugMode {
+			if err != nil {
+				log.Error("Could not send telemetry data", zap.Error(err))
+			} else {
+				log.Info("Telemetry data sent")
+			}
+		}
 	}
 }
 
@@ -188,24 +207,21 @@ func wunderctlBinaryPath() string {
 }
 
 func init() {
-	config.InitConfig()
+	_, isTelemetryDisabled := os.LookupEnv("WG_TELEMETRY_DISABLED")
+	_, isTelemetryDebugEnabled := os.LookupEnv("WG_TELEMETRY_DEBUG")
 
-	viper.SetDefault("OAUTH_CLIENT_ID", "wundergraph-production-cli")
-	viper.SetDefault("OAUTH_BASE_URL", "https://accounts.wundergraph.com/auth/realms/master")
-	viper.SetDefault("API_URL", "https://api.wundergraph.com")
+	config.InitConfig(!isTelemetryDisabled)
+
+	// Can be overwritten by WG_API_URL=<url> env variable
+	viper.SetDefault("API_URL", "https://gateway.wundergraph.com")
 
 	rootCmd.PersistentFlags().StringVarP(&rootFlags.CliLogLevel, "cli-log-level", "l", "info", "sets the CLI log level")
 	rootCmd.PersistentFlags().StringVarP(&DotEnvFile, "env", "e", ".env", "allows you to set environment variables from an env file")
 	rootCmd.PersistentFlags().BoolVar(&rootFlags.DebugMode, "debug", false, "enables the debug mode so that all requests and responses will be logged")
+	rootCmd.PersistentFlags().BoolVar(&rootFlags.Telemetry, "telemetry", !isTelemetryDisabled, "enables telemetry. Telemetry allows us to accurately gauge WunderGraph feature usage, pain points, and customization across all users.")
+	rootCmd.PersistentFlags().BoolVar(&rootFlags.TelemetryDebugMode, "telemetry-debug", isTelemetryDebugEnabled, "enables the debug mode for telemetry. Understand what telemetry is being sent to us.")
 	rootCmd.PersistentFlags().BoolVar(&rootFlags.PrettyLogs, "pretty-logging", false, "switches to human readable format")
 	rootCmd.PersistentFlags().StringVar(&_wunderGraphDirConfig, "wundergraph-dir", files.WunderGraphDirName, "path to your .wundergraph directory")
 	rootCmd.PersistentFlags().BoolVar(&disableCache, "no-cache", false, "disables local caches")
 	rootCmd.PersistentFlags().BoolVar(&clearCache, "clear-cache", false, "clears local caches during startup")
-}
-
-func authenticator() *auth.Authenticator {
-	return auth.New(log,
-		viper.GetString("OAUTH_BASE_URL"),
-		viper.GetString("OAUTH_CLIENT_ID"),
-	)
 }