+++ docs-website/config/navigation.js
@@ -14,6 +14,10 @@ const navigation = [
 				title: 'Manifesto',
 				href: '/manifesto',
 			},
+			{
+				title: 'Usage Information',
+				href: '/usage-information',
+			},
 		],
 	},
 	{
@@ -116,6 +120,10 @@ const navigation = [
 				href: '/docs/guides/enable-autocompletion-in-your-ide',
 			},
 			{
+				title: 'Testing',
+				href: '/docs/guides/testing',
+			},
+			{
 				title: 'Debugging',
 				href: '/docs/guides/debugging',
 			},
@@ -196,6 +204,10 @@ const navigation = [
 				href: '/docs/examples/nextjs',
 			},
 			{
+				title: 'NextJS + React Query',
+				href: '/docs/examples/nextjs-react-query',
+			},
+			{
 				title: 'Vite + SWR',
 				href: '/docs/examples/vite-swr',
 			},
@@ -206,6 +218,10 @@ const navigation = [
 			{
 				title: 'Caching',
 				href: '/docs/examples/caching',
+			},
+			{
+				title: 'Expo + SWR',
+				href: '/docs/examples/expo-swr',
 			},
 			{
 				title: 'FaunaDB NextJS',
@@ -266,6 +282,10 @@ const navigation = [
 			{
 				title: 'WunderGraph schema extension',
 				href: '/docs/examples/schema-extension',
+			},
+			{
+				title: 'Next.js Todos',
+				href: '/docs/examples/nextjs-todos',
 			},
 		],
 	},
@@ -455,7 +475,7 @@ const navigation = [
 				href: '/docs/supported-frontend-frameworks',
 			},
 			{
-				title: 'React-JS',
+				title: 'React',
 				href: '/docs/supported-frontend-frameworks/react-js',
 			},
 			{
@@ -463,7 +483,7 @@ const navigation = [
 				href: '/docs/supported-frontend-frameworks/react-native',
 			},
 			{
-				title: 'NextJS',
+				title: 'Next.js',
 				href: '/docs/supported-frontend-frameworks/nextjs',
 			},
 			{
@@ -926,6 +946,27 @@ const navigation = [
 			{
 				title: 'Overview',
 				href: '/docs/wundergraph-manifest-json-reference',
+			},
+		],
+	},
+	{
+		title: 'Clients reference',
+		links: [
+			{
+				title: 'TypeScript Client',
+				href: '/docs/clients-reference/typescript-client',
+			},
+			{
+				title: 'SWR',
+				href: '/docs/clients-reference/swr',
+			},
+			{
+				title: 'React Query',
+				href: '/docs/clients-reference/react-query',
+			},
+			{
+				title: 'Next.js',
+				href: '/docs/clients-reference/nextjs',
 			},
 		],
 	},