+++ testapps/vite/.wundergraph/wundergraph.config.ts
@@ -16,7 +16,7 @@ const spaceX = introspect.graphql({
 
 const weather = introspect.graphql({
 	apiNamespace: 'weather',
-	url: 'https://graphql-weather-api.herokuapp.com/',
+	url: 'https://weather-api.wundergraph.com/',
 });
 
 /*const jsonPlaceholder = introspect.openApi({
@@ -121,7 +121,7 @@ configureWunderGraphApplication({
 	],
 	codeGenerators: [
 		{
-			templates: [...templates.typescript.all, templates.typescript.operations, templates.typescript.linkBuilder],
+			templates: [...templates.typescript.all],
 		},
 		{
 			templates: [templates.typescript.client],