+++ packages/create-wundergraph-app/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.1](https://github.com/wundergraph/wundergraph/compare/create-wundergraph-app@0.2.0...create-wundergraph-app@0.2.1) (2022-12-08)
+
+### Bug Fixes
+
+* use the version from package.json for the --version command ([#409](https://github.com/wundergraph/wundergraph/issues/409)) ([13de661](https://github.com/wundergraph/wundergraph/commit/13de661fd520d2cc165e7e251d1f33bcbde6e8d3)) (@JivusAyrus)
+
+## [0.2.0](https://github.com/wundergraph/wundergraph/compare/create-wundergraph-app@0.1.0...create-wundergraph-app@0.2.0) (2022-12-06)
+
+### Features
+
+* **create-wundergraph-app:** download latest tag instead of main ([#349](https://github.com/wundergraph/wundergraph/issues/349)) ([a1b82a1](https://github.com/wundergraph/wundergraph/commit/a1b82a1204401e538b3341e797e19f8e67b69095)) (@fiam)
+
 ## 0.1.0 (2022-11-18)
 
 ### Features